+++ src/lightning_app/runners/cloud.py
@@ -55,9 +55,9 @@ from lightning_app.core.constants import (
     DEFAULT_NUMBER_OF_EXPOSED_PORTS,
     DISABLE_DEPENDENCY_CACHE,
     ENABLE_APP_COMMENT_COMMAND_EXECUTION,
+    enable_interruptible_works,
     enable_multiple_works_in_default_container,
     ENABLE_MULTIPLE_WORKS_IN_NON_DEFAULT_CONTAINER,
-    enable_preemptible_works,
     ENABLE_PULLING_STATE_ENDPOINT,
     ENABLE_PUSHING_STATE_ENDPOINT,
     get_cloud_queue_type,
@@ -542,7 +542,7 @@ class CloudRuntime(Runtime):
                 name=work.cloud_compute.name,
                 count=1,
                 disk_size=work.cloud_compute.disk_size,
-                preemptible=work.cloud_compute.preemptible,
+                preemptible=work.cloud_compute.interruptible,
                 shm_size=work.cloud_compute.shm_size,
             )
 
@@ -584,7 +584,7 @@ class CloudRuntime(Runtime):
             user_requested_flow_compute_config=V1UserRequestedFlowComputeConfig(
                 name=self.app.flow_cloud_compute.name,
                 shm_size=self.app.flow_cloud_compute.shm_size,
-                preemptible=False,
+                interruptible=False,
             ),
             works=works,
             local_source=True,
@@ -651,11 +651,11 @@ class CloudRuntime(Runtime):
         if get_cluster_driver():
             v1_env_vars.append(V1EnvVar(name="LIGHTNING_CLUSTER_DRIVER", value=get_cluster_driver()))
 
-        if enable_preemptible_works():
+        if enable_interruptible_works():
             v1_env_vars.append(
                 V1EnvVar(
-                    name="LIGHTNING_ENABLE_PREEMPTIBLE_WORKS",
-                    value=os.getenv("LIGHTNING_ENABLE_PREEMPTIBLE_WORKS", "0"),
+                    name="LIGHTNING_ENABLE_interruptible_WORKS",
+                    value=os.getenv("LIGHTNING_ENABLE_interruptible_WORKS", "0"),
                 )
             )