+++ tests/tests_app/structures/test_structures.py
@@ -54,7 +54,7 @@ def test_dict():
                 "mounts": None,
                 "shm_size": 0,
                 "_internal_id": "default",
-                "preemptible": False,
+                "interruptible": False,
             },
         }
         for k in ("a", "b", "c", "d")
@@ -88,7 +88,7 @@ def test_dict():
                 "mounts": None,
                 "shm_size": 0,
                 "_internal_id": "default",
-                "preemptible": False,
+                "interruptible": False,
             },
         }
         for k in ("a", "b", "c", "d")
@@ -122,7 +122,7 @@ def test_dict():
                 "mounts": None,
                 "shm_size": 0,
                 "_internal_id": "default",
-                "preemptible": False,
+                "interruptible": False,
             },
         }
         for k in ("a", "b", "c", "d")
@@ -208,7 +208,7 @@ def test_list():
                 "mounts": None,
                 "shm_size": 0,
                 "_internal_id": "default",
-                "preemptible": False,
+                "interruptible": False,
             },
         }
         for i in range(4)
@@ -242,7 +242,7 @@ def test_list():
                 "mounts": None,
                 "shm_size": 0,
                 "_internal_id": "default",
-                "preemptible": False,
+                "interruptible": False,
             },
         }
         for i in range(4)
@@ -271,7 +271,7 @@ def test_list():
                 "mounts": None,
                 "shm_size": 0,
                 "_internal_id": "default",
-                "preemptible": False,
+                "interruptible": False,
             },
         }
         for i in range(4)