+++ tests/tests_app/utilities/packaging/test_cloud_compute.py
@@ -67,13 +67,13 @@ def test_cloud_compute_clone():
             assert c1_dict[k] == c2_dict[k]
 
 
-def test_preemptible(monkeypatch):
-    """Test preemptible can be enabled with env variables and for GPU only."""
+def test_interruptible(monkeypatch):
+    """Test interruptible can be enabled with env variables and for GPU only."""
     with pytest.raises(ValueError, match="isn't supported yet"):
-        CloudCompute("gpu", preemptible=True)
+        CloudCompute("gpu", interruptible=True)
 
-    monkeypatch.setenv("LIGHTNING_ENABLE_PREEMPTIBLE_WORKS", "1")
+    monkeypatch.setenv("LIGHTNING_ENABLE_interruptible_WORKS", "1")
     with pytest.raises(ValueError, match="supported only with GPU"):
-        CloudCompute("cpu", preemptible=True)
+        CloudCompute("cpu", interruptible=True)
 
-    CloudCompute("gpu", preemptible=True)
+    CloudCompute("gpu", interruptible=True)